implement -forkid parameter (MVHF-BU-DES-CSIG-1)
Watch out: specifying hexadecimal value e.g. `-forkid=0x666000` will result in value 0 being assigned quietly. The parameter conversion is just broken like that (looks like it's being accomodating to Windows platform lack of string-to-long conversion capable of handling base 16 automatically). 

So stick to decimal values (this needs to be documented as well in case it's not).