Merge PR33 - Triggering adjustments (set FinalActivateForkHeight) and wallet backup fixes, etc.
This PR covers some necessary fixes to the way we handle wallet backup.

Wallet backup is now also checked+done (if necessary) in ActivateFork(). This is done so that the wallet is backed up even if the fork is triggered by SegWit before the fork height or auto backup block height are reached. That's what users will expect - a reliable backup when the fork triggers. This also matches the requirements MVHF-BU-USER-REQ-10 and its children.

The wallet backup check in UpdateTip() remains, so that wallet backups on specific pre-fork blocks can still be done as before. However, the check and possible triggering of ActivateFork() is moved in front of the wallet backup check in UpdateTip(), and if necessary pre-empts it.

The fork activation now sets FinalActivateForkHeight to the actual fork height, so other functions like the difficulty retargeting periods should work correctly even if the fork is triggered by SegWit. The actual fork height + actual backup block height are now also written correctly to btcfork.conf when the fork triggers.

Other changes include adaptation of the wallet backup test to take into account that triggering the fork before the wallet backup block no longer works the same way (therefore the forkheight was shifted to autobackupblock+1.

FUTURE WORK

The wallet backup test may need to be extended with another node test run where the node forkheight < autobackupblock , to check that it backs up correctly at the fork point and not again at the autobackupblock when the fork is already active.

A "backup on SegWit activation" test case should also be included - this should now work due to the wallet backup code being called in ActivateFork().