introduce a fork consensus code id
To prepare for testnet testing, it will be useful to easily identify builds that are compatible / incompatible w.r.t. their post-fork consensus behaviour.

Builds that have the same consensus code should remain on the same chain as long as their other fork parameters are equal (and in the case of MVF-BU, the emergent-consensus related parameters need to be compatible too).