added 6th node to test no soft-fork wallet backup if node is stopped after pre-fork block and restarted later
This just adds the node, it does not yet add the test steps that check that no redundant backup is created on the fork block.

There are also a bunch of minor loop refactorings. The current state of the script is committed to show that it still passes with these refactorings before further code changes.